==> code/main analysis/8_figures.R <==
########################################.
#### Script to generate the main figures in the manuscript, some figures are generated in script #2 though
#### Author:  Gerard Rocher-Ros
#### Last edit: 2023-03-10
########################################.


# Load  and install packages ----
# List of all packages needed
package_list <- c('tidyverse', 'sf',  'ggpubr', 'rnaturalearth', 'ggtext', 'lubridate', 'patchwork', 'see', 'broom', 'readxl')

# Check if there are any packacges missing
packages_missing <- setdiff(package_list, rownames(installed.packages()))

# If we find a package missing, install them
if(length(packages_missing) >= 1) install.packages(packages_missing) 

# Now load all the packages
lapply(package_list, require, character.only = TRUE)

#custom functions ----
# Function to calculate gas saturation, 
# from temperature ( in Kelvin) and atm pressure (derived from elevation (in m.a.s.l) for CH4 
# Henry's Law constants from http://www.henrys-law.org
# Temperature correction using van't Hoff equation , temperature is in (Kelvin)
# Kh is in ("Kh, cp") = (mol/L/Atm) at STP
# Concentrations (mole fraction, also ppmv) in the atmosphere are approximate for 2013, should be updated over time
# AtmP is in (atmospheres)
get_Ch4eq <- function(temperature, elevation){
  
  pressure <- (1-(.0000225577*elevation))^5.25588
  Kh <- (1.4*10^-3)*exp(1700*((1/temperature)-(1/298)))
  AtmosphereConc <-  1.83
  EquilSaturation <- AtmosphereConc*Kh*pressure #umol/L, mmol/m3
  
  return(EquilSaturation)
}



# Load files ----
# aggregated hex layers

meth_hexes_avg <- st_read( "data/processed/GIS/meth_hexes_footprint_corr.shp") %>%
  mutate(Fch4_mean = ifelse(Fch4_mean > .11, NA, Fch4_mean) ) %>% #there is a super tiny hex in Svalvard that gets a high value, due to the aggregation, i remove it
  st_transform("+proj=eqearth +wktext") 

meth_hexes_flux_corr <- st_read( "data/processed/GIS/meth_hexes_flux_corr.shp") %>%
  st_transform("+proj=eqearth +wktext") 

meth_hexes_k_corr <- st_read( "data/processed/GIS/meth_hexes_k_corr.shp") %>%
  st_transform("+proj=eqearth +wktext") 

meth_hexes_uncapped <- st_read( "data/processed/GIS/meth_hexes_uncapped.shp") %>%
  st_transform("+proj=eqearth +wktext") 

ice_cover <- st_read( "data/processed/GIS/icecov_hexes.shp") %>%
  st_transform("+proj=eqearth +wktext") 

extrap_pols <- st_read( "data/processed/GIS/extrap_sites.shp") %>%
  st_transform("+proj=eqearth +wktext") 

mountains <- st_read("data/processed/GIS/mountains/CMEC_Mountains_Enh2018.shp") %>%
  st_transform("+proj=eqearth +wktext") 

meth_fluxes <- read_csv("data/processed/grades_ch4_fluxes.csv", lazy = FALSE) %>% 
  dplyr::select(-contains("sd"))

grades <- read_csv("data/processed/grades_coords.csv") 

grimeDB_attributes <- read_csv("data/processed/grimeDB_concs_with_grade_attributes.csv") %>% 
  drop_na(lat) %>% 
  filter(Aggregated == "No",
         !str_detect(Channel_type,"DD|PI|GT|PS|TH|Th|DIT")) %>%
  mutate( month = month.name[month(date)]) %>%
  dplyr::select(COMID, Site_ID, lat, lon, month) %>% 
  group_by(COMID, month) %>% 
  summarise(lat= first(lat),
            lon = first(lon)) %>% 
  st_as_sf( coords = c("lon", "lat"),  crs = 4326) %>%
  st_transform("+proj=eqearth +wktext") 

#join it into one file 
meth_gis <- meth_fluxes %>% 
  left_join( grades, by = "COMID") 

rm(meth_fluxes, grades)

# get the modelled k values 
# load file with discharge and k data 
hydro_k <- read_csv("data/processed/q_and_k.csv") %>% 
  dplyr::select(COMID, Slope, ends_with("k")) %>% 
  mutate( k_mean = rowMeans(pick(ends_with("k")), na.rm = TRUE))

#do average methane fluxes rates
meth_avg <- meth_gis %>% 
  left_join(hydro_k %>%  dplyr::select(COMID, Slope, k_mean), by = "COMID") %>% 
  drop_na(Jan_ch4F) %>% 
  mutate(Fch4_mean = rowMeans(pick(ends_with("ch4F")), na.rm = TRUE)) %>% 
  dplyr::select(COMID, Slope, k_mean, Fch4_mean )


## Now get GRiMeDB 
load(file.path("data", "GRiMeDB.rda"))

sites_df <- sites_df %>% 
  mutate(Channel_type = ifelse(is.na(Channel_type) == TRUE, "normal", Channel_type)) 

grime_comids <- read_csv("data/processed/sites_meth_comid.csv") 



# join the flux with concentration and the site dataset, also calcute some variables we need for the temperature dependence
# for that we need the boltzmann constant

boltzmann_k <- 8.62e-5

flux_sites <- flux_df %>% 
  left_join( conc_df, by=c("Site_ID", "Date_start"), multiple = "all") %>% 
  left_join( sites_df, by="Site_ID") %>% 
  filter(Aggregated == "No", WaterTemp_degC < 100, 
         !Channel_type %in% c("DD","GT", "PS", "TH","Th", "CAN")) %>%
  mutate(WaterTemp_best = ifelse(is.na(WaterTemp_degC) == TRUE, WaterTemp_degC_estimated, WaterTemp_degC) ) %>% 
  drop_na(Diffusive_CH4_Flux_Mean, WaterTemp_best ) %>% 
  mutate(site_lat = paste("Site:",Site_ID,"\nLat: " , round(Latitude,1), sep=""), #make a new variables that has site id and latitude
         temp_inv = 1/(boltzmann_k*(WaterTemp_best + 273.15)), #inverse of temperature with boltzmann constant
         temp_stand = 1/(boltzmann_k*288.15) - 1/(boltzmann_k*(WaterTemp_best + 273.15)))  #and the same but standrdadized at 15 degC which keeps things more straight

#sort by latitude for later plots
flux_sites$site_lat <- reorder(flux_sites$site_lat, desc(flux_sites$Latitude))

#download world map
world <-  ne_download(scale = 110, type = 'land', category = 'physical', returnclass = "sf") %>%
  st_transform("+proj=eqearth +wktext") 

#and lake layer
lakes <- ne_download(scale = 50, type = 'lakes', category = 'physical', returnclass = "sf") %>%
  st_transform("+proj=eqearth +wktext") 


#buffer df to shrink each hex according to runoff
buffers <- meth_hexes_avg %>% 
  mutate( 
    buffer_change = case_when(
      runoff < .005 ~ -45000,
      runoff >= .005 & runoff < .02 ~ -35000,
      runoff >= .02 & runoff < .04 ~ -25000,
      runoff >= .04 & runoff < .06 ~ -15000,
      runoff >= .06 & runoff < .1 ~ -8000,
      runoff >= .1 & runoff < .5 ~ -6000,
      runoff >= .5 & runoff < 1 ~ -3000,
      runoff >= 1 & runoff < 1.5 ~ -2000,
      runoff >= 1.5 & runoff < 2 ~ -1000,
      runoff >= 2  ~ 0,
      is.na(runoff) == TRUE ~ 0)
  ) %>%
  dplyr::select(buffer_change) %>% 
  mutate( buffer_change = ifelse(is.na(buffer_change) == TRUE, 0, buffer_change)) %>% 
  st_drop_geometry()




# Map figures ----
## map of concentrations 
map_ch4 <- 
  ggplot() +
  geom_sf(
    data = st_buffer(meth_hexes_avg, dist = buffers$buffer_change), color = NA,
    aes( fill = ch4_mean) )+
  geom_sf(data = lakes %>% filter(scalerank < 2), fill="aliceblue", color=NA)+
  scale_fill_viridis_c(
    option = "magma", na.value = "gray",
    direction = - 1,
    name = "**CH<sub>4</sub> concentration** <br>(mmol m<sup>-3</sup>)")+
  coord_sf(xlim = c(-15000000, 16000000), ylim = c(-8600000, 8600000), expand = FALSE) +
  guides( fill = guide_colourbar(title.position = "top"))+
  theme_void()+
  theme(text = element_text(family = "Helvetica"),
        legend.position = c(0.125, 0.35),
        legend.direction = "vertical",
        legend.title = ggtext::element_markdown(size = 8),
        legend.text = element_text(size =7 ),
        legend.key.height  = unit(.35, 'cm'),
        legend.key.width =  unit(.25, 'cm'))


## map of fluxes
map_fch4 <- 
  ggplot()+
  geom_sf(
    data = st_buffer(meth_hexes_avg, dist = buffers$buffer_change), 
    aes(fill = Fch4_mean*16/12), color = NA)+
  geom_sf(data=lakes %>% filter(scalerank < 2), fill="aliceblue", color=NA)+
  scale_fill_viridis_c(
    option = "magma", na.value = "gray",
    direction = -1,
    name = "**CH<sub>4</sub> emission** <br> (g CH<sub>4</sub> m^-2 d<sup>-1</sup>)")+
  coord_sf(xlim = c(-15000000, 16000000), ylim = c(-8600000, 8600000), expand = FALSE) +
  guides( fill = guide_colourbar(title.position = "top"))+
  theme_void()+
  theme(
        text = element_text(family = "Helvetica"),
        legend.position = c(0.1, 0.35),
        legend.direction = "vertical",
        legend.title = ggtext::element_markdown(size = 8),
        legend.text = element_text(size =7 ),
        legend.key.height  = unit(.35, 'cm'),
        legend.key.width =  unit(.25, 'cm'))


#put them together and add labels
maps_combined <- map_ch4 + 
  map_fch4 + 
  plot_annotation(tag_levels = 'a') +
  plot_layout(ncol = 1) & 
  theme(plot.tag.position = c(0.03, 1))

#ggsave(maps_combined, filename = "figures/fig_maps.png", dpi= 1000, scale = .8)

#export publication quality figure 
ggsave(maps_combined, filename = "figures/figure1.pdf", width= 120, height= 130,  units= "mm")


# Map of the ss of concentrations ----
#map of sd of concentrations
map_ch4_sd <- 
ggplot() +
  geom_sf(
    data = meth_hexes_avg, color = NA,
    aes( fill = ch4_sd) )+
  geom_sf(data=lakes %>% filter(scalerank < 2), fill="aliceblue", color=NA)+
  scale_fill_viridis_c(
    option = "magma", na.value = "gray",
    direction = -1,
    #trans = "pseudo_log", 
    name = "**SD of <br> predicted CH<sub>4</sub>** <br>(mmol m<sup>-3</sup>)")+
  guides( fill = guide_colourbar(title.position = "top"))+
  coord_sf(xlim = c(-16000000, 16000000), ylim = c(-7300000, 8600000), expand = FALSE) +
  theme_void()+
  theme(text = element_text(family = "Helvetica"),
        legend.position = c(0.17, 0.35),
        legend.direction = "vertical",
        legend.title = ggtext::element_markdown(size= 9),
        legend.key.height  = unit(.5, 'cm'),
        legend.key.width =  unit(.3, 'cm'))


ggsave(map_ch4_cv, filename = "figures/supplementary/extended_data4.jpeg", width = 183, height = 140, dpi = 300, units = "mm")


# Monthly maps ----

#Show the modelled concentrations by month, with ice cover
labelled_months <- tibble(month= month.abb, labels =month.name)

#prepare the files, pivot to longer for faceting and fix the labels
meth_monthly <- meth_hexes_avg %>% 
  st_drop_geometry() %>% 
  dplyr::select(index, Jan_ch4:Dec_ch4) %>% 
  pivot_longer(-index, values_to = "ch4", names_to = "month") %>% 
  mutate(month = str_remove(month, "_ch4")) %>% 
  left_join(labelled_months, by = "month") %>% 
  mutate(labels = fct_relevel(labels, month.name)) %>% 
  left_join(meth_hexes_avg %>% select(index), by = "index") %>% 
  st_sf() 

ice_monthly <- ice_cover %>% 
  select(-runoff) %>% 
  st_drop_geometry() %>% 
  pivot_longer(ends_with("icecov"), values_to = "ice_frac", names_to = "month") %>% 
  mutate(month = str_remove(month, "_icecov"),
         ice_frac = ifelse(ice_frac > 0.6, 1, NA)) %>% 
  left_join(labelled_months, by = "month") %>% 
  mutate(labels = fct_relevel(labels, month.name)) %>% 
  left_join(ice_cover %>% select(index), by = "index") %>% 
  drop_na(ice_frac) %>% 
  st_sf()

## plot of monthly concentrations ----
map_monthly <- 
  ggplot( meth_monthly) +
  geom_sf(aes(fill = ch4), color = NA )+
  geom_sf(data = lakes %>% filter(scalerank < 2), fill="aliceblue", color=NA)+
  geom_sf(data = ice_monthly, fill= "powderblue", color=NA)+
  scale_fill_viridis_c(
    option = "magma", na.value = "gray",
    direction = -1,
    name = "**CH<sub>4</sub> concentration** (mmol m<sup>-3</sup>)")+
    facet_wrap( ~ labels, ncol = 3)+
  guides( fill = guide_colourbar(title.position = "top"))+
  coord_sf(xlim = c(-13000000, 16000000), ylim = c(-7300000, 8300000), expand = FALSE) +
  theme_void()+
  theme(text = element_text(family = "Helvetica"),
        legend.position = "bottom",
        legend.direction = "horizontal",
        legend.title = ggtext::element_markdown(size = 10),
        legend.key.width =  unit(1, 'cm'),
        legend.key.height =  unit(.3, 'cm'),
        legend.text = element_text(size = 9),
        strip.text = element_text( size = 11))


ggsave(map_monthly, 
       filename = "figures/supplementary/extended_data3.jpeg", width = 183, height = 140, dpi = 900, units = "mm")



## Map of extrapolated areas monthly, with sampled sites ----
#prepare the file, pivot longer for facet_wrap and fix month labels
extrap_pols_monthly <- extrap_pols %>% 
  st_drop_geometry() %>% 
  pivot_longer(-index, values_to = "extrap", names_to = "month") %>% 
  mutate(month = str_remove(month, "interp_"),
         extrap = ifelse(extrap < 0.9, 1, NA)) %>% 
  left_join(labelled_months, by = "month") %>% 
  mutate(labels = fct_relevel(labels, month.name)) %>% 
  left_join(extrap_pols  %>% select(index), by = "index") %>% 
  drop_na(extrap) %>% 
  st_sf()

n_month <- grimeDB_attributes %>% 
  st_drop_geometry() %>% 
  group_by(labels= month) %>% 
  summarise(n= n() ) %>% 
  mutate(labels = fct_relevel(labels, month.name))

map_extrap_areas <- 
  ggplot(extrap_pols_monthly)+
  geom_sf(data = ice_monthly, fill = "powderblue", color = NA)+
  geom_sf(data = world, fill = NA, color = "gray50")+
  geom_sf(fill = "brown4", color = NA)+
  geom_sf(data = grimeDB_attributes %>% 
            mutate(labels = as_factor(month) %>% fct_relevel(month.name)), color="black", size = .5)+
  geom_text( data = n_month,
             mapping = aes(x = 2500000, y = -5500000, label =n), family = "Helvetica", size= 3)+
  coord_sf(xlim = c(-13000000, 16000000), ylim = c(-7600000, 8300000), expand = FALSE)+
  facet_wrap( ~ labels, ncol = 3)+
  theme_void()+
  theme(strip.text = element_text( size = 11))

ggsave(map_extrap_areas, 
       filename = "figures/supplementary/extended_data1.jpeg", width = 183, height = 140, dpi = 500, units = "mm")



# Figure with seasonality emissions along latitudinal bands ----

# Prepare file with latitudinal bands, and sum emissions for that 
seasonal_df <- meth_gis %>% 
  mutate(lat_label = case_when(lat >= 70 ~ " > 70 \u00B0N",
                               lat >= 60 & lat < 70 ~ "60 - 70 \u00B0N",
                               lat >= 50 & lat < 60 ~ "50 - 60 \u00B0N",
                               lat >= 40 & lat < 50 ~ "40 - 50 \u00B0N",
                               lat >= 30 & lat < 40 ~ "30 - 40 \u00B0N",
                               lat >= 20 & lat < 30 ~ "20 - 30 \u00B0N",
                               lat >= 10 & lat < 20 ~ "10 - 20 \u00B0N",
                               lat >=  0 & lat < 10 ~ "0 - 10 \u00B0N",
                               lat >= -10 & lat < 0 ~ "0 - 10 \u00B0S",
                               lat >= -20 & lat < -10 ~ "10 - 20 \u00B0S",
                               lat >= -30 & lat < -20 ~ "20 - 30 \u00B0S",
                               lat >= -40 & lat < -30 ~ "30 - 40 \u00B0S",
                               lat >= -30 & lat < -20 ~ "20 - 30 \u00B0S",
                               lat >= -40 & lat < -30 ~ "30 - 40 \u00B0S",
                               lat >= -50 & lat < -40 ~ "40 - 50 \u00B0S",
                               lat < -50 ~ " > 50 \u00B0S") ) %>% 
  group_by(lat_label) %>% 
  summarise(lat = mean(lat),
            Jan_ch4E = sum(Jan_ch4E_reach + Jan_ch4E_extrap, na.rm = TRUE)/1e+12*16/12, #in Mg Ch4
            Feb_ch4E = sum(Feb_ch4E_reach + Feb_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Mar_ch4E = sum(Mar_ch4E_reach + Mar_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Apr_ch4E = sum(Apr_ch4E_reach + Apr_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            May_ch4E = sum(May_ch4E_reach + May_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Jun_ch4E = sum(Jun_ch4E_reach + Jun_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Jul_ch4E = sum(Jul_ch4E_reach + Jul_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Aug_ch4E = sum(Aug_ch4E_reach + Aug_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Sep_ch4E = sum(Sep_ch4E_reach + Sep_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Oct_ch4E = sum(Oct_ch4E_reach + Oct_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Nov_ch4E = sum(Nov_ch4E_reach + Nov_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Dec_ch4E = sum(Dec_ch4E_reach + Dec_ch4E_extrap, na.rm = TRUE)/1e+12*16/12,
            Jan_ch4F = mean(Jan_ch4F, na.rm = TRUE),
            Feb_ch4F = mean(Feb_ch4F, na.rm = TRUE),
            Mar_ch4F = mean(Mar_ch4F, na.rm = TRUE),
            Apr_ch4F = mean(Apr_ch4F, na.rm = TRUE),
            May_ch4F = mean(May_ch4F, na.rm = TRUE),
            Jun_ch4F = mean(Jun_ch4F, na.rm = TRUE),
            Jul_ch4F = mean(Jul_ch4F, na.rm = TRUE),
            Aug_ch4F = mean(Aug_ch4F, na.rm = TRUE),
            Sep_ch4F = mean(Sep_ch4F, na.rm = TRUE),
            Oct_ch4F = mean(Oct_ch4F, na.rm = TRUE),
            Nov_ch4F = mean(Nov_ch4F, na.rm = TRUE),
            Dec_ch4F = mean(Dec_ch4F, na.rm = TRUE),
            Jan_areakm2 = sum(Jan_area_m2, na.rm = TRUE)/1e+6,
            Feb_areakm2 = sum(Feb_area_m2, na.rm = TRUE)/1e+6,
            Mar_areakm2 = sum(Mar_area_m2, na.rm = TRUE)/1e+6,
            Apr_areakm2 = sum(Apr_area_m2, na.rm = TRUE)/1e+6,
            May_areakm2 = sum(May_area_m2, na.rm = TRUE)/1e+6,
            Jun_areakm2 = sum(Jun_area_m2, na.rm = TRUE)/1e+6,
            Jul_areakm2 = sum(Jul_area_m2, na.rm = TRUE)/1e+6,
            Aug_areakm2 = sum(Aug_area_m2, na.rm = TRUE)/1e+6,
            Sep_areakm2 = sum(Sep_area_m2, na.rm = TRUE)/1e+6,
            Oct_areakm2 = sum(Oct_area_m2, na.rm = TRUE)/1e+6,
            Nov_areakm2 = sum(Nov_area_m2, na.rm = TRUE)/1e+6,
            Dec_areakm2 = sum(Dec_area_m2, na.rm = TRUE)/1e+6)

#pivot to longer for faceting
seasonal_df_long <- seasonal_df %>% 
  pivot_longer(-c(lat, lat_label), names_to = c("month", "parameter"), names_pattern = "(.*)_(.*)",
               values_to = "value") %>% 
  mutate(lat_label= fct_reorder(as_factor(lat_label), lat,  .desc = TRUE),
         month = fct_relevel(as_factor(month), month.abb)) %>% 
  pivot_wider(names_from = parameter, values_from = value)

#make the plot of monthly emissions by latitudinal bands
season_lat_plot <- 
  seasonal_df_long %>% 
  ggplot(aes(month, ch4E, fill=areakm2))+
  geom_col()+
  scale_y_continuous(trans = "sqrt", breaks = c(0, 0.1, 0.3), 
                     limits = c(0, NA) , expand=c(0,0))+
  scale_fill_viridis_c(trans="log10", direction = -1)+
  facet_wrap(~lat_label, ncol = 1, strip.position = "right")+
  theme_bw()+
  labs(x = "", y = "**CH<sub>4</sub> emissions** <br> (Tg CH<sub>4</sub>)" , fill="**River area** <br> (km^2 )")+
  theme(axis.ticks.length.x = unit(0, "cm"),
        strip.background = element_blank(),#element_rect(color=NA, fill = "gray80"),
        strip.text.y = element_blank(), #element_text(size = 12, angle=0),
        axis.text.x = element_text(size = 6, color = "gray5"),
        axis.text.y = element_text(size= 4, color = "gray40"),
        panel.border =  element_blank(), 
        panel.grid = element_blank(),
        axis.line.y = element_line(color="gray50", linewidth = .2),
        axis.ticks.y = element_line(color="gray50", linewidth = .2),
        axis.title.y = ggtext::element_markdown(size = 8),
        legend.title = ggtext::element_markdown(size = 7),
        legend.text = element_text(size = 5),
        legend.position = c(.5,-.15), legend.direction = "horizontal",
        legend.key.width = unit(7, "mm"), legend.key.height = unit( 2, "mm"),
        plot.margin = unit(c(0,0,1,0), "mm")
  )

#make a plot of total emissions by latitudanl bands
total_lats <- 
  seasonal_df_long %>% 
  group_by(lat_label) %>% 
  summarise(lat= mean(lat), 
            total_flux= sum(ch4E)) %>% 
  mutate(lat_label = fct_reorder(as_factor(lat_label), lat,  .desc = FALSE) ) %>% 
  ggplot(aes( x= total_flux, y= lat_label))+
  geom_col(fill="gray80")+
  #geom_text(aes(x=2, label=lat_label))+
  scale_x_continuous(expand = c(0,0), breaks = c(0,1,2))+
  labs(x = "**CH<sub>4</sub> emissions** <br> (Tg CH<sub>4</sub>)", y="")+
  theme_classic()+
  theme(axis.ticks.length.y = unit(0, "mm"),
        #axis.text.y = element_blank(),
        axis.text.y = element_text(size=6, color="black"),
        axis.text.x = element_text(size=6),
        axis.line.x = element_line(color="gray40", linewidth = .2),
        axis.ticks.x = element_line(color="gray40", linewidth = .2),
        axis.line.y = element_blank(),
        axis.title.x = ggtext::element_markdown(size = 8, color="black"),
        panel.background = element_rect(fill = "transparent"), # bg of the panel
        plot.background = element_rect(fill = "transparent", color = NA), # bg of the plot
        plot.margin = unit(c(0,0,0,0), "mm")
  )

# put them together and save them
#figure_2 <- season_lat_plot +  inset_element(total_lats, left = .97, right = 1.4, bottom = -.118, top = 1.02)
figure_2 <- season_lat_plot + total_lats + plot_layout(ncol = 2, widths = c(3, 1))

figure_2
#ggsave("figures/emissions_lat_season.png",figure_2, width =8, height = 8, dpi = 800)

#export of production quality figures 
ggsave( "figures/figure2.pdf", figure_2,  width =120, height = 85, units = "mm")


## Figure about the footprint correction of the flux rates -----
#Map alternatives of flux estimates, with capping k, capping with fluxes, correcting k by footprint and uncorrected
## map of fluxes
map_footprint <- 
  ggplot()+
  geom_sf(
    data = st_buffer(meth_hexes_avg,  dist = buffers$buffer_change), 
    aes(fill = Fch4_mean*16/12), color = NA )+
  geom_sf(data=lakes %>% filter(scalerank < 2), fill="aliceblue", color=NA)+
  geom_sf(data = mountains, fill = NA, color= "gray50", size= .2)+
 # annotate("text", x = 3000000, y = -6000000, label = expression("Total flux = 13.4 Tg "~CH[4]~yr^-1 ), size= 3 )+
  scale_fill_viridis_c(
    option = "magma", na.value = "gray",
    direction = -1, breaks= c(0, 0.05, 0.1),
    name = "**CH<sub>4</sub> flux** (g CH<sub>4</sub> m^-2 d<sup>-1</sup>)")+
  coord_sf(xlim = c(-15000000, 16000000), ylim = c(-8600000, 8600000), expand = FALSE) +
  guides( fill = guide_colourbar(title.position = "top"))+
 # labs(title = "Reach footprint correction")+
  theme_void()+
  theme(legend.position = c(0.5, -0.2),
        legend.direction = "horizontal",
        legend.title = ggtext::element_markdown(size = 8),
        legend.key.height  = unit(.2, 'cm'),
        legend.key.width =  unit(.65, 'cm'),
        plot.title = element_text(hjust = .5))

map_k_capped <- 
  ggplot()+
  geom_sf(
    data = st_buffer(meth_hexes_k_corr,  dist = buffers$buffer_change), 
    aes(fill = Fch4_mean*16/12), color = NA )+
  geom_sf(data=lakes %>% filter(scalerank < 2), fill="aliceblue", color=NA)+
  geom_sf(data = mountains, fill = NA, color= "gray50", size= .2)+
 # annotate("text", x = 3000000, y = -6000000, label = expression("Total flux = 11.8 Tg "~CH[4]~yr^-1 ), size= 3 )+
  scale_fill_viridis_c(
    option = "magma", na.value = "gray",
    direction = -1, breaks= c(0, 0.2, 0.4),
    name = "**CH<sub>4</sub> flux** (g CH<sub>4</sub> m^-2 d<sup>-1</sup>)")+
  coord_sf(xlim = c(-15000000, 16000000), ylim = c(-8600000, 8600000), expand = FALSE) +
  guides( fill = guide_colourbar(title.position = "top"))+
 #labs(title = "k capped")+
  theme_void()+
  theme(legend.position = c(0.5, -0.2),
        legend.direction = "horizontal",
        legend.title = ggtext::element_markdown(size = 8),
        legend.key.height  = unit(.2, 'cm'),
        legend.key.width =  unit(.65, 'cm'),
        plot.title = element_text(hjust = .5)) 


map_flux_capped <- 
  ggplot()+
  geom_sf(
    data = st_buffer(meth_hexes_flux_corr,  dist = buffers$buffer_change), 
    aes(fill = Fch4_mean*16/12), color = NA )+
  geom_sf(data=lakes %>% filter(scalerank < 2), fill="aliceblue", color=NA)+
  geom_sf(data = mountains, fill = NA, color= "gray50", size= .2)+
  #annotate("text", x = 3000000, y = -6000000, label = expression("Total flux = 14.5 Tg "~CH[4]~yr^-1 ), size= 3 )+
  scale_fill_viridis_c(
    option = "magma", na.value = "gray",
    direction = -1, breaks= c(0, 0.05, 0.1),
    name = "**CH<sub>4</sub> flux** (g CH<sub>4</sub> m^-2 d<sup>-1</sup>)")+
  coord_sf(xlim = c(-15000000, 16000000), ylim = c(-8600000, 8600000), expand = FALSE) +
  guides( fill = guide_colourbar(title.position = "top"))+
  #labs(title= "Fluxes capped")+
  theme_void()+
  theme(legend.position = c(0.5, -0.2),
        legend.direction = "horizontal",
        legend.title = ggtext::element_markdown(size = 8),
        legend.key.height  = unit(.2, 'cm'),
        legend.key.width =  unit(.65, 'cm'),
        plot.title = element_text(hjust = .5))

map_uncapped <- 
  ggplot()+
  geom_sf(
    data = st_buffer(meth_hexes_uncapped,  dist = buffers$buffer_change), 
    aes(fill = Fch4_mean*16/12), color = NA )+
  geom_sf(data=lakes %>% filter(scalerank < 2), fill="aliceblue", color=NA)+
  geom_sf(data = mountains, fill = NA, color= "gray50", size= .2)+
 # annotate("text", x = 3000000, y = -6000000, label = expression("Total flux = 14.7 Tg "~CH[4]~yr^-1 ), size= 3 )+
  scale_fill_viridis_c(
    option = "magma", na.value = "gray",
    direction = -1, breaks= c(0, 0.5, 1),
    name = "**CH<sub>4</sub> flux** (g CH<sub>4</sub> m^-2 d<sup>-1</sup>)")+
  coord_sf(xlim = c(-15000000, 16000000), ylim = c(-8600000, 8600000), expand = FALSE) +
  guides( fill = guide_colourbar(title.position = "top"))+
  #labs(title= "No correction")+
  theme_void()+
  theme(legend.position = c(0.5, -0.2),
        legend.direction = "horizontal",
        legend.title = ggtext::element_markdown(size = 8),
        legend.key.height  = unit(.2, 'cm'),
        legend.key.width =  unit(.65, 'cm'),
        plot.title = element_text(hjust = .5))



#now look at how data compares with real obs
#read file to compare the two methods for upscaling 
methods_comp <- read_csv("data/processed/methods_flux_comparison.csv") %>% 
  mutate(method =  as_factor(method) %>% 
           fct_recode( `Fluxes capped` = "flux_cap", `k capped` = "k_cap",
                       `Reach footprint correction` = "footprint", Uncorrected = "uncorrected") %>% 
           fct_relevel("Uncorrected", "k capped", "Fluxes capped", "Reach footprint correction" ))


#density plot comparison
density_uncorrected <- ggplot()+
  stat_density_2d(data= flux_sites %>% filter(CH4mean> 0, Diffusive_CH4_Flux_Mean > 0), 
                  geom = "polygon", contour = TRUE,
                  aes(CH4mean, Diffusive_CH4_Flux_Mean, alpha = ..level..), 
                  breaks = c(0.05, 0.1, 0.25, 0.5, 1, 1.5, 2))+
  stat_density_2d(data = slice_sample(methods_comp %>% filter(method == "Uncorrected"), n= 5e+6), 
                  geom = "polygon", aes(x=ch4, y=flux*1000/16, alpha = ..level..),
                  breaks = c(0.05, 0.1, 0.25, 0.5, 1, 1.5, 2),
                  fill = "red4", color= "brown", linewidth = .1)+
  scale_x_log10()+
  scale_y_log10(labels = scales::number)+
  labs(x= "", y = "")+
  facet_wrap(~method)+
  theme_bw()+
  theme(legend.position =  "none",
        axis.text = element_text(size= 8),
        strip.background = element_blank(),
        strip.text = element_text(size=10, face = "bold"))

density_flux <- ggplot()+
  stat_density_2d(data= flux_sites %>% filter(CH4mean> 0, Diffusive_CH4_Flux_Mean > 0), 
                  geom = "polygon", contour = TRUE,
                  aes(CH4mean, Diffusive_CH4_Flux_Mean, alpha = ..level..), 
                  breaks = c(0.05, 0.1, 0.25, 0.5, 1, 1.5, 2))+
  stat_density_2d(data = slice_sample(methods_comp %>% filter(method == "Fluxes capped"), n= 5e+6), 
                  geom = "polygon", aes(x=ch4, y=flux*1000/16, alpha = ..level..),
                  breaks = c(0.05, 0.1, 0.25, 0.5, 1, 1.5, 2),
                  fill = "red4", color= "brown", linewidth = .1)+
  scale_x_log10()+
  scale_y_log10(labels = scales::number)+
  facet_wrap(~method)+
  labs(x= "",  y = "CH<sub>4</sub> flux (mmol m^-2 d<sup>-1</sup>)")+
  theme_bw()+
  theme(legend.position =  "none",
        axis.text = element_text(size= 8),
        axis.title.y = element_markdown(size=9, hjust=-3.5),
        strip.background = element_blank(),
        strip.text = element_text(size=10, face = "bold"))

density_k <- ggplot()+
  stat_density_2d(data= flux_sites %>% filter(CH4mean> 0, Diffusive_CH4_Flux_Mean > 0), 
                  geom = "polygon", contour = TRUE,
                  aes(CH4mean, Diffusive_CH4_Flux_Mean, alpha = ..level..), 
                  breaks = c(0.05, 0.1, 0.25, 0.5, 1, 1.5, 2))+
  stat_density_2d(data = slice_sample(methods_comp %>% filter(method == "k capped"), n= 5e+6), 
                  geom = "polygon", aes(x=ch4, y=flux*1000/16, alpha = ..level..),
                  breaks = c(0.05, 0.1, 0.25, 0.5, 1, 1.5, 2),
                  fill = "red4", color= "brown", linewidth = .1)+
  scale_x_log10()+
  scale_y_log10(labels = scales::number)+
  facet_wrap(~method)+
  labs(x= "",  y = "")+
  theme_bw()+
  theme(legend.position =  "none",
        axis.text = element_text(size= 8),
        axis.title.y = element_markdown(size=9, hjust=3.5),
        strip.background = element_blank(),
        strip.text = element_text(size=10, face = "bold"))

density_footprint <- ggplot()+
  stat_density_2d(data= flux_sites %>% filter(CH4mean> 0, Diffusive_CH4_Flux_Mean > 0), 
                  geom = "polygon", contour = TRUE,
                  aes(CH4mean, Diffusive_CH4_Flux_Mean, alpha = ..level..), 
                  breaks = c(0.05, 0.1, 0.25, 0.5, 1, 1.5, 2))+
  stat_density_2d(data = slice_sample(methods_comp %>% filter(method == "Reach footprint correction"), n= 5e+6), 
                  geom = "polygon", aes(x=ch4, y=flux*1000/16, alpha = ..level..),
                  breaks = c(0.05, 0.1, 0.25, 0.5, 1, 1.5, 2),
                  fill = "red4", color= "brown", linewidth = .1)+
  scale_x_log10()+
  scale_y_log10(labels = scales::number)+
  facet_wrap(~method)+
  labs(x="CH<sub>4</sub> concentration (mmol m<sup>-3</sup>)", y="",
       caption = "Grey contours are empirical data from GRiMeDB")+
  theme_bw()+
  theme(legend.position =  "none",
        axis.text = element_text(size= 8),
        axis.title.x = element_markdown(size=9),
        strip.background = element_blank(),
        strip.text = element_text(size=10, face = "bold"))






plot_corrections <- density_uncorrected + map_uncapped +
  density_k + map_k_capped +
  density_flux + map_flux_capped + 
  density_footprint + map_footprint + 
  plot_layout(ncol = 2, widths = 1) +
  plot_annotation(tag_levels = 'a') & 
  theme(plot.tag.position = c(0.07, .93))




ggsave("figures/supplementary/extended_data5.jpeg", plot_corrections,  width = 150, height = 200, dpi = 500, units = "mm")


# Comparison of the modelled fluxes with measured fluxes in GRiMeDB data ----
# Process all files ----
sites_clean <- sites_df %>% 
  left_join(grime_comids, by="Site_ID") %>% 
  drop_na(COMID)



## attach the COMID to the concentration df and keep useful variables 
conc_df_comids <- conc_df %>% 
  filter(Site_ID %in% sites_clean$Site_ID) %>% 
  left_join(sites_clean, by="Site_ID") %>% 
  dplyr::select(Site_ID, Source_ID = Source_ID.x , Aggregated, Channel_type, COMID,
                Latitude, distance_snapped, CH4mean, CO2mean,
                date= Date_start, discharge_measured= Q, WaterTemp_degC, WaterTemp_degC_estimated  ) %>% 
  mutate(CH4mean =ifelse(CH4mean < 0, 0.0001, CH4mean)) %>% 
  drop_na(CH4mean)

#Look at sites with highest concentrations
conc_df_comids %>% 
  group_by(Site_ID) %>% 
  summarise(CH4mean= median(CH4mean),
            Latitude = first(Latitude),
            Site_ID = first(Site_ID),
            COMID = first(COMID)) %>% 
  arrange(desc(CH4mean)) %>% 
  dplyr::select(CH4mean, Latitude, Site_ID, COMID) %>% 
  print(n=30)


#get the flux df and fix some issues, also merge with conc_Df
flux_comid <-  flux_df %>% 
  mutate(Diffusive_CH4_Flux_Mean =ifelse(Diffusive_CH4_Flux_Mean < 0, 0, Diffusive_CH4_Flux_Mean)) %>% 
  filter(Site_ID %in% sites_clean$Site_ID) %>% 
  left_join(sites_clean, by="Site_ID") %>% 
  left_join(conc_df, by=c("Site_ID", "Date_start"), multiple = "all") %>% 
  mutate(WaterTemp_best = if_else(is.na(WaterTemp_degC) == FALSE, WaterTemp_degC, WaterTemp_degC_estimated )) #unify temperatures in one column

#now join the observed fluxes and predicted into one df. I will filter the sites that are not well snapped (>500m)
#this is to compare modelled with obswrved fluxes
flux_comp <- flux_comid %>% 
  dplyr::select(COMID, Site_ID, date= Date_start, distance_snapped, CH4mean, WaterTemp_best, elev=Elevation_m, 
                CO2mean, Diffusive_CH4_Flux_Mean, Diff_Method, k_Method ) %>% 
  filter(distance_snapped < 500) %>% 
  left_join( meth_gis, by="COMID") %>%  
  left_join(hydro_k, by = "COMID") %>% 
  drop_na(Diffusive_CH4_Flux_Mean) %>% 
  mutate(CH4mean_ex = CH4mean - get_Ch4eq(WaterTemp_best + 273.15, elev),
         CH4mean_ex = ifelse(is.na(CH4mean_ex) == TRUE,
                             CH4mean - get_Ch4eq(20+ 273.15, elev),
                             CH4mean_ex),
         k_obs =Diffusive_CH4_Flux_Mean/CH4mean,  # estimate empirical kch4 (m d-1) as the quotient of flux (mmol m-2 d-1) and concentration (mmol m-3)
         month_obs= month(date),
         k_pred =  case_when(month_obs == 1 ~  Jan_k,
                             month_obs == 2 ~  Feb_k,
                             month_obs == 3 ~  Mar_k,
                             month_obs == 4 ~  Apr_k,
                             month_obs == 5 ~  May_k,
                             month_obs == 6 ~  Jun_k,
                             month_obs == 7 ~  Jul_k,
                             month_obs == 8 ~  Aug_k,
                             month_obs == 9 ~  Sep_k,
                             month_obs == 10 ~ Oct_k,
                             month_obs == 11 ~ Nov_k,
                             month_obs == 12 ~ Dec_k),
         flux_pred = case_when(month_obs == 1 ~  Jan_ch4F,
                               month_obs == 2 ~  Feb_ch4F,
                               month_obs == 3 ~  Mar_ch4F,
                               month_obs == 4 ~  Apr_ch4F,
                               month_obs == 5 ~  May_ch4F,
                               month_obs == 6 ~  Jun_ch4F,
                               month_obs == 7 ~  Jul_ch4F,
                               month_obs == 8 ~  Aug_ch4F,
                               month_obs == 9 ~  Sep_ch4F,
                               month_obs == 10 ~ Oct_ch4F,
                               month_obs == 11 ~ Nov_ch4F,
                               month_obs == 12 ~ Dec_ch4F),
         sc_obs =  1824 - 98.12 * WaterTemp_best + 2.413 * (WaterTemp_best^2) - 0.0241 * (WaterTemp_best^3), #calculate the schmidt number
         k600_obs = k_obs*(600/sc_obs)^(-.5), #so estimate the k600 observed
         flux_pred = flux_pred*1000/12, #and the preducted flux in the same units
         k600_pred = k_pred*(600/sc_obs)^(-.5) # and the k600 predicted
         ) %>% 
  group_by(COMID) %>% #now for every COMID calculate median values 
  summarise(k600_pred = median(k600_pred),
            k600_obs = median(k600_obs),
            flux_pred = median(flux_pred) ,
            flux_obs = median(Diffusive_CH4_Flux_Mean),
            k_Method = first(k_Method)) %>% 
  filter( !k_Method %in% c("not determined", "other"), flux_pred > 0, flux_obs > 0) #remove negative fluxes and strange methods for k


# Figure to compare modelled fluxes with measured methane fluxes. This is all values
plot_overall <- ggplot(data= flux_comp,
       aes(flux_obs, flux_pred))+
  geom_point( color="black", size=2)+
  geom_abline(slope=1, intercept = 0, linetype=2)+
  geom_smooth(method="lm", se= FALSE, color= "red", linetype = 2)+
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~")))+
  scale_x_log10(labels=scales::number, limits = c(0.0001, 1000))+
  scale_y_log10(labels=scales::number, limits = c(0.0001, 1000))+
  labs(x="Measured CH<sub>4</sub> diffusive emissions", y="Modelled CH<sub>4</sub> diffusive emissions", title = "All data")+
  theme_bw()+
  theme(legend.key.size = unit(1, 'cm'), 
        axis.title.x = ggtext::element_markdown(),
        axis.title.y = ggtext::element_markdown())

#As we are comparing very different things here, now we do the same comparison for sites and occasions when k was +/- 50% similar betwen modeleld and observed
plot_zoomed <- ggplot(data= flux_comp %>% 
         filter( !k_Method %in% c("not determined", "other"),
                  k600_pred/k600_obs > .5 &  k600_pred/k600_obs < 1.5, 
                flux_pred > 0, flux_obs > 0),
         aes(flux_obs, flux_pred))+
    geom_point(data=flux_comp %>% filter( !k_Method %in% c("not determined", "other"),
                                          flux_pred > 0, flux_obs > 0, is.na(k600_obs) == FALSE, k600_pred/k600_obs < 100),
               aes(flux_obs, flux_pred, color = (k600_pred/k600_obs)*100), alpha=.5, size=2)+
  geom_point( color="black", size=2)+
  geom_abline(slope=1, intercept = 0, linetype=2)+
  geom_smooth(method="lm", color= "red")+
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~")), label.y.npc = .99)+
  stat_regline_equation(label.y.npc = .9)+
  scale_x_log10(labels=scales::number, limits = c(0.01, 100))+
  scale_y_log10(labels=scales::number, limits = c(0.1, 10))+
  scale_color_viridis_c(trans="log10", labels=scales::number, breaks = c(10, 50,  150, 1000, 10000))+
  labs(x = "Measured CH<sub>4</sub> diffusive emissions", y="", 
       color = "% difference\n predicted \n vs observed k",
       title = "Points with comparable k<sub>600</sub> (+/- 50%)")+
  theme_bw()+
  theme(legend.key.size = unit(1, 'cm'),
        axis.title.x = ggtext::element_markdown(),
        plot.title =  ggtext::element_markdown())

#put them together and save them
plot_overall + 
  plot_zoomed +
  plot_annotation(tag_levels = 'a') & 
  theme(plot.tag.position = c(0.05, 1), plot.tag = element_text(size=15))

ggsave("figures/supplementary/extended_data9.jpeg", width = 180, height = 80, units = "mm", scale= 1.4)



# Temperature dependence of diffusive fluxes ----
# We start by plotting the temperature emission relationship of all sites that have n> 20, coloured by latitude

#mycolors <- colorRampPalette(RColorBrewer::brewer.pal(8, "Set1"))(21)

#plot with all sites with a n > 20
plot_sites <- 
  flux_sites %>% 
  filter(Diffusive_CH4_Flux_Mean > 0.001) %>% 
  ggplot(aes(temp_stand, Diffusive_CH4_Flux_Mean, color = abs(Latitude)))+
  geom_point( size = 2, alpha = .2)+
  geom_smooth(data= flux_sites %>% 
                filter(Diffusive_CH4_Flux_Mean > 0.001) %>% 
                add_count(Site_ID) %>% 
                filter(n > 20) %>% 
                arrange(abs(Latitude)),  
              method= "lm", se= FALSE, size = .8,
              aes(temp_stand, Diffusive_CH4_Flux_Mean, group = Site_ID), alpha=.6)+
  geom_smooth(method= "lm", color="gray20", se= FALSE, size = 1)+
  geom_abline(slope=.96, intercept = -.39, linetype = 2)+
  annotate("text", x=2.8, y = 360, label="m = 0.96", angle = 33)+
  scale_color_viridis_c(option = "viridis", name = "Latitude")+
  scale_x_continuous(name= expression(paste("Standardized temperature ", bgroup("(", frac(1, kT) - frac(1, kT[C]), ")" ))), limits= c(-2.6, 3.1),
                     sec.axis = sec_axis(~ (273.15*boltzmann_k* . + 0.0520562)/(0.00347041 - boltzmann_k* . ), name =expression("Temperature " ( degree*C)), 
                                         breaks = seq(0, 40, by=5)) )+ # solved using wolfram alpha, y= 1/(k*(288.15)) - 1/(k*(x+273.15)), solve x 
  scale_y_continuous(trans= "log10", labels = scales::number, breaks = c(.01, .1, 1, 10, 100),
                     name= "Diffusive CH<sub>4</sub> emissions (mmol m^-2 d<sup>-1</sup>)")+
  theme_classic()+
  theme(axis.title.y = ggtext::element_markdown(), legend.position = c(.1,.8), legend.key.height = unit(4, 'mm'),
        legend.background=element_blank())



# Comparison with other aquatic systems ----
#we use the suppplementary data from Yvon-durocher et al. 2014. Can be downloaded here: https://doi.org/10.1038/nature13164 
# we keep sites with more than 5 observations, and remove the ones without published data
yvon_durocher2014 <- read_excel("~/Desktop/yvon-durocher2014.xlsx") %>% 
  filter(!temp == "see notes") %>% 
  mutate(across(c(temp, flux), as.numeric),
         flux_mmol.m2 = flux/16,
         ecosystem.type = recode(ecosystem.type, "Aquatic" = "Lakes", "Rice Paddy" = "Rice Paddies", "Wetland" = "Wetlands")) %>% 
  add_count(site) %>%
  filter(n > 5) %>% 
  dplyr::select(site, latitude, longitude, ecosystem.type, temp_c = temp, flux_mmol.m2)

#reformot our data and put together with the other aquatic systems
data_for_Em <- flux_sites  %>% 
  filter(Diffusive_CH4_Flux_Mean > 0, WaterTemp_degC > -30) %>% 
  add_count(Site_ID) %>% 
  filter(n > 5) %>% 
  mutate(ecosystem.type = "Rivers",
         site = as.character(Site_ID)) %>% 
  dplyr::select(site, latitude = Latitude, longitude = Longitude, ecosystem.type, 
                temp_c = WaterTemp_degC, flux_mmol.m2 = Diffusive_CH4_Flux_Mean) %>% 
  bind_rows(yvon_durocher2014) %>% 
  mutate(temp_inv = 1/(boltzmann_k*(temp_c + 273.15)))
  
#estimate Ems using linear models, need to try with lme
Em_all <- data_for_Em %>% 
  group_by(site) %>% 
  do(fit = lm(log(flux_mmol.m2) ~ temp_inv, data = .)) %>% 
  summarize(site = first(site),
            tidy(fit))  %>% 
  left_join(data_for_Em %>% 
              group_by(site) %>% 
              summarise(across(c(ecosystem.type, latitude, longitude), first))) 


# do the density plot of Em across systems
comp_systems <- Em_all %>% 
  filter(term == "temp_inv") %>% 
  mutate(ecosystem.type = ecosystem.type %>% 
           fct_relevel("Rivers", after =Inf) %>%
           fct_relevel("Lakes", after= 2)) %>% 
  filter(estimate < 5) %>% #there is an outlier in rivers that we remove
  ggplot()+
  geom_density(aes(-estimate, fill= ecosystem.type, color = ecosystem.type), alpha =.2, size= 1) +
  geom_vline(xintercept = 0, linetype = 2)+ 
  scale_fill_manual(values = c("olivedrab", "mediumseagreen", "cornflowerblue", "darkblue"))+
  scale_color_manual(values = c("olivedrab", "mediumseagreen", "cornflowerblue", "darkblue"))+
  scale_x_continuous(breaks = c(-3, -2, -1, 0, 1, 2, 3), expand = c(0,0))+
  scale_y_continuous(expand = c(0,0))+
  theme_classic()+
  theme(legend.position = c(.2, .86))+
  guides(color = guide_legend(ncol = 1),
         fill = guide_legend(ncol = 1))+
  labs(x= "Apparent activation energy (eV)", y = "Kernel density", color= "", fill = "")


plot_sites +
  comp_systems +
  plot_layout(ncol = 1) +
  plot_annotation(tag_levels = 'a') & 
  theme( plot.tag = element_text(size = 15))


#ggsave("figures/temperature_methane_rivers.png", width = 5, height = 7.5)
ggsave("figures/figure4.pdf", width = 89, height = 133, units = "mm", scale= 1.45)

Em_all %>% 
  filter(term == "temp_inv") %>% 
  group_by(ecosystem.type) %>% 
  summarise(mean(-estimate),
            median(-estimate))
Em_all %>% 
  filter(term == "temp_inv", ecosystem.type == "Rivers") %>% 
  summarise(quantile( -estimate, 0.25),
            quantile(-estimate, 0.75))

em_all_data <-flux_sites %>%
  filter(Diffusive_CH4_Flux_Mean > 0, WaterTemp_degC > -30) %>% 
  do(fit = lm(log(Diffusive_CH4_Flux_Mean) ~ temp_inv, data = .)) 
  
em_all_data$fit

wilcox.test(Em_all %>% 
              filter(term == "temp_inv", ecosystem.type == "Rivers") %>% 
              pull(estimate), 
            Em_all %>% 
              filter(term == "temp_inv", ecosystem.type == "Lakes") %>% 
              pull(estimate))

wilcox.test(Em_all %>% 
              filter(term == "temp_inv", ecosystem.type == "Rivers") %>% 
              pull(estimate), 
            Em_all %>% 
              filter(term == "temp_inv", ecosystem.type == "Wetlands") %>% 
              pull(estimate))

wilcox.test(Em_all %>% 
              filter(term == "temp_inv", ecosystem.type == "Rivers") %>% 
              pull(estimate), 
            Em_all %>% 
              filter(term == "temp_inv", ecosystem.type == "Rice Paddies") %>% 
              pull(estimate))

wilcox.test(Em_all %>% 
              filter(term == "temp_inv", ecosystem.type == "Wetlands") %>% 
              pull(estimate), 
            Em_all %>% 
              filter(term == "temp_inv", ecosystem.type == "Lakes") %>% 
              pull(estimate))




ebullition_temp <- flux_sites %>% 
  filter(Eb_CH4_Flux_Mean > 0.001) %>% 
  ggplot(aes(temp_stand, (Eb_CH4_Flux_Mean)))+
  geom_point(aes( color=abs(Latitude)),  size = 3, alpha = .6)+
  geom_smooth(method= "lm", color="gray20", se= FALSE, size=2)+
  #stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~")), label.y.npc = .99)+
  stat_regline_equation(label.y.npc = .99)+
  scale_color_viridis_c(name="Latitude")+
  annotate("text", x=-2.1, y = 100, label="n = 245")+
  scale_x_continuous(name= expression(paste("Standardized temperature ", bgroup("(", frac(1, kT) - frac(1, kT[C]), ")" ))), limits= c(-2.6, 3.1),
                     sec.axis = sec_axis(~ (273.15*boltzmann_k* . + 0.0520562)/(0.00347041 - boltzmann_k* . ), name =expression("Temperature " ( degree*C)), 
                                         breaks = seq(0, 40, by=5)) )+ # solved using wolfram alpha, y= 1/(k*(288.15)) - 1/(k*(x+273.15)), solve x 
  scale_y_continuous(trans= "log10", labels = scales::number, breaks = c(.01, .1, 1, 10, 100),
                     name= "Ebullitive CH<sub>4</sub> emissions <br> (mmol m^-2 d<sup>-1</sup>)")+
  theme_classic()+
  theme(axis.title.y = ggtext::element_markdown(), legend.position = "right")

ggsave("figures/supplementary/extended_data6.jpeg", ebullition_temp,  scale = .55)


## Temperature dependence of ebullition and across river size ----
flux_sites %>% 
  filter(Diffusive_CH4_Flux_Mean > 0.001) %>% 
  left_join(grime_comids %>% select(Site_ID, order)) %>%
  drop_na(order) %>% 
  #drop_na(System_size) %>% 
 # mutate(System_size = fct_relevel(System_size, "large", after = Inf)) %>% 
  ggplot(aes(temp_stand, Diffusive_CH4_Flux_Mean))+
  geom_point(aes( color=abs(Latitude)),  size = 2, alpha = .6)+
  geom_smooth(method= "lm", color="gray20", se= FALSE, size = 1)+
  stat_regline_equation(label.y.npc = .99)+
  scale_color_viridis_c(name="Latitude", limits= c(0, 71))+
  #geom_abline(slope = .96, intercept = -.39, linetype = 2)+
  #annotate("text", x = 2.7, y = 90, label="m = 0.96", angle = 48)+
  #annotate("text", x = -2, y = 250, label="n = 4,417")+
  scale_x_continuous(name= expression(paste("Standardized temperature ", bgroup("(", frac(1, kT) - frac(1, kT[C]), ")" ))), limits= c(-2.6, 3.1),
                     sec.axis = sec_axis(~ (273.15*boltzmann_k* . + 0.0520562)/(0.00347041 - boltzmann_k* . ), name =expression("Temperature " ( degree*C)), 
                                         breaks = seq(0, 40, by=5)) )+ # solved using wolfram alpha, y= 1/(k*(288.15)) - 1/(k*(x+273.15)), solve x 
  scale_y_continuous(trans= "log10", labels = scales::number, breaks = c(.01, .1, 1, 10, 100),
                     name= "Diffusive CH<sub>4</sub> flux <br> (mmol m^-2 d<sup>-1</sup>)")+
  theme_classic()+
  theme(axis.title.y = ggtext::element_markdown(), legend.position = "right",
        strip.text = element_text(size=14, face= "bold"))+
  facet_wrap(~order)

ggsave("figures/supplementary/river_size_temp.png", width = 12, scale = .7)


# Ebullition assessment ----

mycolors2 <- colorRampPalette(RColorBrewer::brewer.pal(8, "Paired"))(93)
#comparison of ebullition and difussive fluxes in GRIMeDB, we remove observations close to 0, virtually 0
reg_compl_plot <- 
  flux_comid %>% 
  filter(Eb_CH4_Flux_Mean > 0.00001,
         Diffusive_CH4_Flux_Mean > 0.0001) %>% 
  filter(k_Method %in% c("chamber+conc", "tracer addition")) %>% 
  ggplot(aes(Diffusive_CH4_Flux_Mean, Eb_CH4_Flux_Mean ))+
  geom_point( aes( color=as.character(Site_ID)), size = 2, alpha=.7)+
  geom_smooth( method="lm",  color="black", size=1)+
  geom_abline(intercept = 0, slope = 1, linetype=2)+
  stat_cor(aes(label = paste(..rr.label.., ..p.label.., sep = "~`,`~")), label.y.npc = .99)+
  stat_regline_equation(label.y.npc = .9)+
  scale_color_manual(values = mycolors2)+
  scale_y_log10(labels=scales::number)+
  scale_x_log10(labels=scales::number)+
  labs(x = "Diffusive CH<sub>4</sub> emissions <br>(mmol m^-2 d<sup>-1</sup>)", 
       y = "Ebulitive CH<sub>4</sub> emissions <br>(mmol m^-2 d<sup>-1</sup>)")+
  theme_classic()+
  theme(legend.position = "none",
        axis.title.y = ggtext::element_markdown(),
        axis.title.x = ggtext::element_markdown())


#ebullition and difussion density plots
density_comp_plot <- 
  flux_comid %>% 
  filter(Eb_CH4_Flux_Mean > 0.0001,
         Diffusive_CH4_Flux_Mean > 0.0001) %>% 
  filter(k_Method %in% c("chamber+conc", "tracer addition")) %>% 
  pivot_longer(c(Eb_CH4_Flux_Mean, Diffusive_CH4_Flux_Mean ), names_to = "type", values_to = "flux") %>% 
  mutate(type= case_when(type == "Eb_CH4_Flux_Mean" ~ "Ebullition",
                         type == "Diffusive_CH4_Flux_Mean" ~ "Diffusion" )) %>% 
  ggplot(aes(type, flux, fill=type))+
  geom_violindot(fill_dots = "gray30", color_dots="gray30", dots_size = 1, width = 1.5, alpha = .7) +
  stat_summary(fun = "median", geom = "point",  size=4, colour = "black", 
               position = position_nudge(x = 0.1, y = 0), show.legend = FALSE)+
  scale_y_log10(labels=scales::number)+
  scale_fill_manual(values= c("darkblue", "cornflowerblue"))+
  labs(x = "", y = "CH<sub>4</sub> emissions<br> (mmol m^-2 d<sup>-1</sup>)")+
  theme_classic()+
  theme(legend.position = "none",
        axis.title.y = ggtext::element_markdown(),
        axis.ticks.x = element_blank(),
        axis.text.x = element_text(size=11, color="black"))

density_comp_plot + 
  reg_compl_plot +
  plot_annotation(tag_levels = 'a') & 
  theme(plot.tag.position = c(0.1, 1), plot.tag = element_text(size=15))

ggsave("figures/supplementary/extended_data7.jpeg", width = 9, height = 4)






## Figure about the uncertainity and sensitivity analysis -----
## Also calculate the uncertianity on the ebullition estimate

#file with the output of the MC simulations
vals_all <- read_csv("data/processed//uncertaintiy_mc.csv") 

#quick plot looking at the distribution of the uncertainty MC
vals_all %>% filter(type == "normal" ) %>% 
  ggplot(aes(flux))+
  geom_histogram(bins= 60)+
  theme_bw()+
  labs(x= "Flux estimate (Pg CH4 yr-1)", y= "")

#Calculate relevant quantiles of the uncertainity analysis

bg_estimate <- vals_all %>% 
  filter(type == "normal" ) %>% 
  summarise(q.025 = quantile(flux, 0.025),
            q.125 = quantile(flux, 0.125),
            q.25 = quantile(flux, 0.25),
            q.5 = quantile(flux, 0.5),
            q.75 = quantile(flux, 0.75),
            q.875 = quantile(flux, 0.875),
            q.975 = quantile(flux, 0.975)) 

#do a linear  model between log-diff and log-ebullition of observed fluxes in GRiMeDB
#Also, as before, remove obs with negative fluxes and unclear k methods
lm_ebulltion <- flux_comid %>% 
  filter(Eb_CH4_Flux_Mean > 0.00001,
         Diffusive_CH4_Flux_Mean > 0.0001) %>% 
  filter(k_Method %in% c("chamber+conc", "tracer addition")) %>% 
  mutate(diff_log = log(Diffusive_CH4_Flux_Mean),
         eb_log = log(Eb_CH4_Flux_Mean), .keep = "none") %>% 
  lm(eb_log~diff_log, data = .)

# select the 95% interval of the MC of global diffusive estimates. These are the lower and upper bounds
vals_estimate <- data.frame(vals = c("lower", "mid", "upper"),
                            diff_log = c(bg_estimate$q.025, bg_estimate$q.5, bg_estimate$q.975))

#Now use the lm to estimate the corresponding global ebullitive flux, with the prediction intervals, for 
# both the median, lower and upper bounds.
eb_preds <-  predict(lm_ebulltion, vals_estimate, interval = "predict")

#the uncertiainty in the ebullitive estimate is the lower bound of the prediciton interval of the lower bound, and upper of the upper
# this is to include the uncertainity both in the diffusive estimate as well as in the regression between diffusive and ebullitve fluxes
eb_preds

#Finally, do a figure on the sensitivity analysis. needs some serious wrangling for
vals_all %>% 
  mutate(type = fct_recode(type, "**CH<sub>4</sub>**" = "ch4","**River <br>area**" = "area", "**Gas <br>transfer<br> velocity**" = "k"),
         change = fct_recode(change, "+1 SD" = "plus_1SD", "-1 SD" = "minus_1SD")) %>% 
  filter(!type == "normal") %>% 
  group_by(type, change) %>% # first get some useful quantiles for plotting
  summarise( q.025 = quantile(flux, 0.025),
             q.125 = quantile(flux, 0.125),
             q.25 = quantile(flux, 0.25),
             q.5 = quantile(flux, 0.5),
             q.75 = quantile(flux, 0.75),
             q.875 = quantile(flux, 0.875),
             q.975 = quantile(flux, 0.975)) %>% 
  ungroup() %>% 
  ggplot()+
  geom_hline(yintercept = bg_estimate$q.5, size= 2)+ #vertical line for the central estimate
  geom_rect(data = bg_estimate, aes(ymin = q.25, ymax = q.75, xmin = -Inf, xmax = Inf), #grey shadows for increasing quantiles
            fill = "gray30", alpha = 0.3)+
  geom_rect(data = bg_estimate, aes(ymin = q.125, ymax = q.875, xmin = -Inf, xmax = Inf), 
            fill = "gray50", alpha = 0.3)+
  geom_rect(data = bg_estimate, aes(ymin = q.025, ymax = q.975, xmin = -Inf, xmax = Inf), 
            fill = "gray70", alpha = 0.3)+
  geom_point(data= vals_all %>%  #plot all values as thin ines for the sensitivity analysis
               mutate(type = fct_recode(type, "**CH<sub>4</sub>**" = "ch4", "**River <br>area**" = "area", "**Gas <br>transfer<br> velocity**" = "k"),
                      change = fct_recode(change, "+1 SD" = "plus_1SD", "-1 SD" = "minus_1SD")) %>% 
               filter(!type == "normal"),
             aes(x = type, y = flux, color= change),  position = position_dodge(width = .3), alpha= .3, shape= 108, size= 1.5)+
  geom_point(data= vals_all %>% filter(type == "normal"), #same for the main estimate, but put it in the bottom
             aes(x = .43, y = flux),  alpha= .3, shape= 108, size= 3, color = "gray40")+
  geom_point(aes(x = type,  color = change, y= q.5), position = position_dodge(width = .3), size= 6, shape=108, show.legend = F)+ #sensitivity analysis central value
  geom_linerange(aes(x = type,  color = change, ymin = q.25, ymax = q.75), position = position_dodge(width = .3), size= 4)+ # and quantiles, decreasing in size
  geom_linerange(aes(x = type,  color = change, ymin = q.125, ymax = q.875), position = position_dodge(width = .3), size= 3, show.legend = F)+
  geom_linerange(aes(x = type,  color = change, ymin = q.025, ymax = q.975), position = position_dodge(width = .3), size= 2, show.legend = F)+
  scale_color_manual(values = c("brown3", "blue3"))+
  theme_bw()+
  coord_flip()+
  labs(y= "Emission estimate (Pg CH<sub>4</sub> yr<sup>-1</sup>)", x= "", color = "")+
  theme(legend.position = "top", axis.text.y = ggtext::element_markdown(size= 11, hjust = 0),
        axis.title.x = ggtext::element_markdown())

ggsave("figures/supplementary/extended_data8.jpeg", scale= .6)





## checking within reach variability



data_methane_used <- read_csv("data/processed/grimeDB_concs_with_grade_attributes.csv") %>% 
  filter(Aggregated == "No", distance_snapped < 1000,
         !str_detect(Channel_type,"DD|PI|GT|PS|TH|Th|DIT")) %>% 
  mutate(month = month(date,label = TRUE, abbr = TRUE ),
         label =paste0(COMID, "\n (", round(lat,1), "; ", round(lon,1), ")"))

data_methane_used %>% 
  filter(CH4mean > 0.001) %>% 
  add_count(COMID) %>% 
  filter(n>100) %>% 
  mutate(label= as_factor(label) %>% fct_reorder(CH4mean, .fun = mean, .desc= TRUE)) %>% 
  ggplot(aes( label, CH4mean, color= month))+
  geom_jitter(alpha=.2, position = position_dodge(width = .75))+
 # stat_summary(geom = "pointrange", fun.data= "mean_sd", size= 1, color= "black")+
  stat_summary(geom= "pointrange", fun.data= "mean_sd", position = position_dodge(width = .75),
               color= "black", shape = 21, size= .75, aes(fill = month))+
  scale_color_social_d(name = "")+
  scale_fill_social_d(name = "")+
  scale_y_log10(labels=scales::number)+
  labs(x= "COMID \n (lat; lon)", y= expression(CH[4]~(mmol~~m^-3)))+
  theme_classic()+
  theme(legend.position = "top")+
  guides(color = guide_legend(nrow = 1))

ggsave("figures/supplementary/sites_with_many_obs2.png")

 

## figures done for the review process.

conc_sites_df <- grime_comids %>% 
  left_join(sites_df) %>% 
  mutate(lat_classes = case_when(abs(lat) >= 60 ~ "High latitude",
                                 abs(lat) > 25 & abs(lat) < 60 ~ "Mid latitude",
                                 abs(lat) < 25 ~ "Low latitude") %>% 
           fct_relevel("Low latitude", after = 3)) %>% 
  left_join(conc_df %>%  filter(CH4mean > 0.001) %>% summarise(ch4mean = mean(CH4mean), .by = Site_ID))

names(conc_sites_df)

plot_dist <- conc_sites_df %>% 
  group_by(order, lat_classes) %>% 
  summarise(n= n()) %>% 
  ungroup() %>% 
  group_by(lat_classes) %>% 
  mutate(frac_order = n/sum(n)*100) %>% 
ggplot(aes(order,frac_order, fill= lat_classes))+
  geom_col( position = "dodge")+
  scale_x_continuous(breaks= 1:9)+
  scale_y_continuous(name = "Relative distribution (%)", expand = c(0,0))+
  scale_fill_brewer(type= "qual", palette = 3, name = "")+
  theme_bw()+
  theme(legend.position = "top")

plot_concs <- conc_sites_df %>% 
  ggplot(aes(as.character(order), ch4mean))+
  geom_boxplot(outlier.shape = NA)+
  geom_jitter(alpha=.1)+
  scale_y_log10()+
  facet_wrap(~lat_classes, ncol = 1)+
  theme_bw()+
  labs(x= "Order", y = "CH4 concentration (mmol/L)")

plot_dist / plot_concs


#temperature dependence across orders 
#reformot our data and put together with the other aquatic systems
data_for_Em <- flux_sites  %>% 
  left_join(grime_comids %>% select(Site_ID, order)) %>%
  drop_na(order) %>% 
  filter(Diffusive_CH4_Flux_Mean > 0, WaterTemp_degC > -30) %>% 
  add_count(Site_ID) %>% 
  filter(n > 10) %>% 
  mutate(ecosystem.type = "Rivers",
         site = as.character(Site_ID)) %>% 
  dplyr::select(site, latitude = Latitude, longitude = Longitude, ecosystem.type, order, Elevation_m,
                temp_c = WaterTemp_degC, flux_mmol.m2 = Diffusive_CH4_Flux_Mean) %>% 
  bind_rows(yvon_durocher2014) %>% 
  mutate(temp_inv = 1/(boltzmann_k*(temp_c + 273.15)))

temp_dep_size_lat <- data_for_Em %>% 
  group_by(site) %>% 
  do(fit = lm(log(flux_mmol.m2) ~ temp_inv, data = .)) %>% 
  summarize(site = first(site),
            tidy(fit))  %>% 
  left_join(data_for_Em %>% 
              group_by(site) %>% 
              summarise(across(c(ecosystem.type, latitude, longitude, order, Elevation_m), first))) 

temp_dep_size_lat %>% 
  filter(term == "temp_inv", ecosystem.type == "Rivers") %>% 
  ggplot(aes(as.character(order), estimate))+
  geom_boxplot(outlier.shape = NA)+
  geom_hline(yintercept = 0)+
  geom_jitter(alpha=.4)+
  labs(y= "Apparent activation energy (eV)", x = "Order")+
  theme_bw() +

temp_dep_size_lat %>% 
  filter(term == "temp_inv", ecosystem.type == "Rivers", Elevation_m < 1000) %>% 
  ggplot(aes(Elevation_m, estimate))+
  geom_point()+
  labs(y= "Apparent activation energy (eV)", x = "Elevation (m)")+
  theme_bw()
